requirements.txt
Django
djangorestframework
black